issue#6 - fixed changes requested from code review for pr#7. 1) Refractor code to now read in [config-file] via CLI flag instead of using a new Cobra subcommand (As per https://github.com/skycoin/skywire-updater/pull/7#discussion_r275936202). Previous issue of not being able to accept flag arguments due to addition of subcommand was resolved by enabling Command.TraverseChildren. Enabling TraverseChildren allows Cobra to parse local flags on each command before executing target subcommands. 2) Moved /internal/pathutil to /pkg/util/pathutil
