+++ cmd/skywire-updater/commands/gen-config.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/skycoin/skywire-updater/internal/pathutil"
 	"github.com/skycoin/skywire-updater/pkg/update"
+	"github.com/skycoin/skywire-updater/pkg/util/pathutil"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
 )
