+++ go.mod
@@ -3,8 +3,8 @@ module github.com/skycoin/skywire-updater
 require (
 	github.com/FiloSottile/vendorcheck v0.0.0-20160511012517-d6d54d1b5894 // indirect
 	github.com/go-chi/chi v4.0.2+incompatible
-	github.com/golangci/go-tools v0.0.0-20180109140146-af6baa5dc196 // indirect
-	github.com/golangci/golangci-lint v1.15.0 // indirect
+	github.com/golangci/golangci-lint v1.16.0 // indirect
+	github.com/golangci/govet v0.0.0-20180818181408-44ddbe260190 // indirect
 	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
 	github.com/mattn/go-colorable v0.1.1 // indirect
 	github.com/mattn/go-isatty v0.0.6 // indirect
@@ -14,9 +14,6 @@ require (
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.3 // indirect
 	github.com/stretchr/testify v1.3.0
-	golang.org/x/crypto v0.0.0-20190313024323-a1f597ede03a // indirect
-	golang.org/x/net v0.0.0-20190313220215-9f648a60d977 // indirect
-	golang.org/x/sys v0.0.0-20190312061237-fead79001313 // indirect
 	golang.org/x/tools v0.0.0-20190417005754-4ca4b55e2050 // indirect
 	gopkg.in/yaml.v2 v2.2.2
 )
