+++ pkg/update/config.go
@@ -9,7 +9,7 @@ import (
 
 	"gopkg.in/yaml.v2"
 
-	"github.com/skycoin/skywire-updater/internal/pathutil"
+	"github.com/skycoin/skywire-updater/pkg/util/pathutil"
 )
 
 // Config represents an updater service configuration
